refactor: 각 userInfo data를 확장함수 toUserInfo() 처리
issue: #59
