+++ presentation/src/main/java/buy/coke/zet/presentation/intro/splash/SplashActivity.kt
@@ -10,11 +10,11 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.lifecycleScope
 import buy.coke.zet.domain.ServiceResult
-import buy.coke.zet.domain.entitiy.login.LoginResponseEntity
-import buy.coke.zet.presentation.LoginStatus
 import buy.coke.zet.presentation.R
 import buy.coke.zet.presentation.databinding.ActivitySplashBinding
 import buy.coke.zet.presentation.intro.signup.SignUpActivity
+import buy.coke.zet.presentation.model.LoginStatus
+import buy.coke.zet.presentation.model.toUserInfo
 import buy.coke.zet.presentation.product.list.ProductListActivity
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
@@ -33,7 +33,7 @@ class SplashActivity : AppCompatActivity() {
 
             if (autoLoginResult is ServiceResult.Success) {
                 Toast.makeText(this@SplashActivity, "로그인 성공", Toast.LENGTH_SHORT).show()
-//                LoginStatus.userInfo = autoLoginResult.data
+                LoginStatus.userInfo = autoLoginResult.data.toUserInfo()
                 
                 Handler(Looper.getMainLooper()).postDelayed({
                     startActivity(Intent(this@SplashActivity, ProductListActivity::class.java))
